Use self-test.sabnzbd.org for connection tests
