+++ sabnzbd/interface.py
@@ -27,6 +27,7 @@ import urllib
 import json
 import re
 import hashlib
+import socket
 from random import randint
 from xml.sax.saxutils import escape
 
@@ -1504,7 +1505,7 @@ SPECIAL_BOOL_LIST = \
 SPECIAL_VALUE_LIST = \
     ('size_limit', 'folder_max_length', 'fsys_type', 'movie_rename_limit', 'nomedia_marker',
               'req_completion_rate', 'wait_ext_drive', 'history_limit', 'show_sysload', 'ipv6_servers',
-              'rating_host', 'ipv6_test_host'
+              'rating_host', 'selftest_host'
      )
 SPECIAL_LIST_LIST = \
     ('rss_odd_titles', 'prio_sort_list'
@@ -2577,14 +2578,12 @@ class Status(object):
 
         # Dashboard: Begin
         if not kwargs.get('skip_dashboard'):
-            from sabnzbd.utils.getipaddress import localipv4, publicipv4, ipv6
             header['localipv4'] = localipv4()
             header['publicipv4'] = publicipv4()
             header['ipv6'] = ipv6()
             # Dashboard: DNS-check
             try:
-                import socket
-                socket.gethostbyname('www.google.com')
+                socket.gethostbyname(cfg.selftest_host())
                 header['dnslookup'] = "OK"
             except:
                 header['dnslookup'] = None
@@ -3065,7 +3064,6 @@ def rss_history(url, limit=50, search=None):
 
     return rss.write()
 
-
 def rss_warnings():
     """ Return an RSS feed with last warnings/errors """
     rss = RSS()
@@ -3082,3 +3080,36 @@ def rss_warnings():
     rss.channel.lastBuildDate = std_time(time.time())
     rss.channel.pubDate = rss.channel.lastBuildDate
     return rss.write()
+
+def localipv4():
+    try:
+        s_ipv4 = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+        s_ipv4.connect(('1.2.3.4', 80))    # Option: use 100.64.1.1 (IANA-Reserved IPv4 Prefix for Shared Address Space)
+        ipv4 = s_ipv4.getsockname()[0]
+        s_ipv4.close()
+    except:
+        ipv4 = None
+        pass
+    return ipv4
+
+def publicipv4():
+    try:
+        import urllib2
+        req = urllib2.Request("http://" + cfg.selftest_host(), headers={'User-Agent': 'SABnzbd+/%s' % sabnzbd.version.__version__}) 
+        f = urllib2.urlopen(req, timeout=2)    # timeout 2 seconds, in case website is not accessible
+        public_ipv4 = f.read()
+        socket.inet_aton(public_ipv4)  # if we got anything else than a plain IPv4 address, this will raise an exception
+    except:
+        public_ipv4 = None
+        pass
+    return public_ipv4
+
+def ipv6():
+    try:
+        s_ipv6 = socket.socket(socket.AF_INET6, socket.SOCK_DGRAM)
+        s_ipv6.connect(('2001:db8::8080', 80))    # IPv6 prefix for documentation purpose
+        ipv6 = s_ipv6.getsockname()[0]
+        s_ipv6.close()
+    except:
+        ipv6 = None
+    return ipv6
